Vision-GA - vision_localize_objects & vision_localize_object_gcs
