+++ samples/detect.v1p3beta1.js
@@ -17,71 +17,6 @@
 
 'use strict';
 
-function localizeObjects(fileName) {
-  // [START vision_localize_objects_beta]
-  // Imports the Google Cloud client libraries
-  const vision = require('@google-cloud/vision').v1p3beta1;
-  const fs = require('fs');
-
-  // Creates a client
-  const client = new vision.ImageAnnotatorClient();
-
-  /**
-   * TODO(developer): Uncomment the following line before running the sample.
-   */
-  // const fileName = `/path/to/localImage.png`;
-
-  const request = {
-    image: {content: fs.readFileSync(fileName)},
-  };
-
-  client
-    .objectLocalization(request)
-    .then(results => {
-      const objects = results[0].localizedObjectAnnotations;
-      objects.forEach(object => {
-        console.log(`Name: ${object.name}`);
-        console.log(`Confidence: ${object.score}`);
-        const veritices = object.boundingPoly.normalizedVertices;
-        veritices.forEach(v => console.log(`x: ${v.x}, y:${v.y}`));
-      });
-    })
-    .catch(err => {
-      console.error('ERROR:', err);
-    });
-  // [END vision_localize_objects_beta]
-}
-
-function localizeObjectsGCS(uri) {
-  // [START vision_localize_objects_gcs_beta]
-  // Imports the Google Cloud client libraries
-  const vision = require('@google-cloud/vision').v1p3beta1;
-
-  // Creates a client
-  const client = new vision.ImageAnnotatorClient();
-
-  /**
-   * TODO(developer): Uncomment the following line before running the sample.
-   */
-  // const uri = `gs://bucket/bucketImage.png`;
-
-  client
-    .objectLocalization(uri)
-    .then(results => {
-      const objects = results[0].localizedObjectAnnotations;
-      objects.forEach(object => {
-        console.log(`Name: ${object.name}`);
-        console.log(`Confidence: ${object.score}`);
-        const veritices = object.boundingPoly.normalizedVertices;
-        veritices.forEach(v => console.log(`x: ${v.x}, y:${v.y}`));
-      });
-    })
-    .catch(err => {
-      console.error('ERROR:', err);
-    });
-  // [END vision_localize_objects_gcs_beta]
-}
-
 function detectHandwritingOCR(fileName) {
   // [START vision_handwritten_ocr_beta]
   // Imports the Google Cloud client libraries
@@ -154,18 +89,6 @@ function detectHandwritingGCS(uri) {
 require(`yargs`)
   .demand(1)
   .command(
-    `localizeObjects`,
-    `Detects Objects in a local image file`,
-    {},
-    opts => localizeObjects(opts.fileName)
-  )
-  .command(
-    `localizeObjectsGCS`,
-    `Detects Objects Google Cloud Storage Bucket`,
-    {},
-    opts => localizeObjectsGCS(opts.gcsUri)
-  )
-  .command(
     `detectHandwriting`,
     `Detects handwriting in a local image file.`,
     {},
@@ -178,19 +101,6 @@ require(`yargs`)
     opts => detectHandwritingGCS(opts.handwritingSample)
   )
   .options({
-    fileName: {
-      alias: 'f',
-      default: `./resources/duck_and_truck.jpg`,
-      global: true,
-      requiresArg: true,
-      type: 'string',
-    },
-    gcsUri: {
-      alias: 'u',
-      global: true,
-      requiresArg: true,
-      type: 'string',
-    },
     handwritingSample: {
       alias: 'h',
       default: `./resources/handwritten.jpg`,
@@ -206,8 +116,6 @@ require(`yargs`)
       type: 'string',
     },
   })
-  .example(`node $0 localizeObjects -f ./resources/duck_and_truck.jpg`)
-  .example(`node $0 localizeObjectsGCS gs://my-bucket/image.jpg`)
   .example(`node $0 detectHandwriting ./resources/handwritten.jpg`)
   .example(`node $0 detectHandwritingGCS gs://my-bucket/image.jpg`)
   .wrap(120)
